Agent/CloudFunctions/requirements.txt
google-cloud-monitoring
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http